Minimalistic Kernel
Skipped building the bootloader, used bootimage instead (https://github.com/rust-osdev/bootimage) but a native bootloader is planned.

Currently using BIOS instead of UEFI but UEFI is planned to be used in the future!

In its current form if executed the OS prints "Hello World!" to the VGA text buffer